\name{pdfXMP}
\alias{pdfXMP}
\title{
  PDF XMP extractor
}
\description{
  Extract XMP data from a PDF file.
}
\usage{
pdfXMP(file, jabrefOnly = TRUE)
}
\arguments{
  \item{file}{
    path of a PDF file
}
  \item{jabrefOnly}{
    When TRUE, only return XMP data written by Jabref.
  }
}
%\details{}
\value{
  A character vector.
}
\author{
Ronggui HUANG
}
